separate CCombFilterBase from CombFilterIf.h & add kNotImplementedError
