+++ src/CombFilter/CombFilterIf.cpp
@@ -116,76 +116,3 @@ float CCombFilterIf::getParam (FilterParam_t eParam) const
     return m_pCCombFilter->getParam(eParam);
 }
 
-CCombFilterBase::CCombFilterBase(float fMaxDelayLengthInS, float fSampleRateInHz, int iNumChannels)
-: m_fGain(0.F), m_iNumChannels(iNumChannels), m_buffer(nullptr)
-{
-    setDelay(fMaxDelayLengthInS);
-}
-
-bool CCombFilterBase::isInRange(float lower, float upper, float value) {
-    return (lower <= value && value < upper);
-}
-
-Error_t CCombFilterBase::setGain(float fGain) {
-    if (!isInRange(0.F, kfMaxGain, fGain)) {
-        return Error_t::kFunctionInvalidArgsError;
-    }
-    m_fGain = fGain;
-    return Error_t::kNoError;
-}
-
-/*! Update the delay length of filter. This would cause deleting & newing CRingBuffer object. 
-\param fDelay the delay time IN SECOND
-*/
-Error_t CCombFilterBase::setDelay(float fDelay) {
-    if (!isInRange(0.F, kfMaxDelay, fDelay)) {
-        return Error_t::kFunctionInvalidArgsError;
-    }
-	m_fDelay = fDelay;
-	m_iDelayInSample = static_cast<int>(m_fDelay / m_fSampleRateInHz);
-	if (m_buffer != nullptr) {
-		delete m_buffer;
-	}
-	m_buffer = new CRingBuffer<float>(m_iDelayInSample);
-    return Error_t::kNoError;
-}
-
-Error_t CCombFilterBase::setParam(FilterParam_t eParam, float fParamValue) {
-    switch (eParam)
-    {
-    case CCombFilterIf::kParamGain:
-        setGain(fParamValue);
-        break;
-    case CCombFilterIf::kParamDelay:
-        setDelay(fParamValue);
-        break;
-    case CCombFilterIf::kNumFilterParams:
-        return Error_t::kFunctionInvalidArgsError;
-    default:
-        break;
-    }
-    return Error_t::kNoError;
-}
-
-float CCombFilterBase::getGain() const {
-    return m_fGain;
-}
-
-float CCombFilterBase::getDelay() const {
-    return m_fDelay;
-}
-
-float CCombFilterBase::getParam(CCombFilterIf::FilterParam_t eParam) const {
-    switch (eParam)
-    {
-    case CCombFilterIf::kParamGain:
-        return getGain();
-    case CCombFilterIf::kParamDelay:
-        return getDelay();
-    case CCombFilterIf::kNumFilterParams:
-        return -1.F;
-    default:
-        break;
-    }
-    return -1.F;
-}
